# backend/docs/api/property-management-openapi.yaml
openapi: 3.0.3
info:
  title: PropMetrik Property Management API
  description: |
    Property Management API for PropMetrik platform.
    
    ## Features
    - **Property Management**: Properties, units, amenities, and listings
    - **Tenancy Management**: Lease lifecycle, tenant onboarding, renewals
    - **Payment Processing**: Rent collection via Paystack (cards, mobile money)
    - **Maintenance**: Work order submission and status tracking
    - **Lease Templates**: Template CRUD, variable substitution, PDF generation
    - **Tenant Portal**: Self-service for tenants (payments, maintenance, documents)
    
    ## Authentication
    - Admin/Staff endpoints use Keycloak Bearer tokens
    - Tenant Portal uses magic link/OTP-based authentication
    
    ## Payment Modes
    - Test mode: Uses Paystack test keys (no real transactions)
    - Live mode: Uses Paystack live keys (real transactions)
    
  version: 1.0.0
  contact:
    name: PropMetrik Support
    email: support@propmetrik.com
servers:
  - url: https://api.propmetrik.com/v1/pm
    description: Production server
  - url: https://staging-api.propmetrik.com/v1/pm
    description: Staging server
  - url: http://localhost:3001/api/pm
    description: Local development
...
paths:
...
  /maintenance/{requestId}:
    get:
      tags:
        - Maintenance
      summary: Get maintenance request
      operationId: getMaintenanceRequest
      parameters:
        - $ref: '#/components/parameters/MaintenanceRequestId'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceRequestDetail'
      security:
        - bearerAuth: []

    patch:
      tags:
        - Maintenance
      summary: Update maintenance request
      description: Update status, assign technician, add notes
      operationId: updateMaintenanceRequest
      parameters:
        - $ref: '#/components/parameters/MaintenanceRequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMaintenanceRequest'
      responses:
        '200':
          description: Request updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceRequestDetail'
      security:
        - bearerAuth: []

  # ===== LEASE TEMPLATES =====
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak JWT token for admin/staff access
    
    tenantToken:
      type: http
      scheme: bearer
      description: Session token for tenant portal access
  parameters:
...
    MaintenanceRequestId:
      name: requestId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    MaintenanceRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenancyId:
          type: string
          format: uuid
        title:
          type: string
        category:
          type: string
        priority:
          type: string
          enum: [low, medium, high, emergency]
        status:
          type: string
          enum: [pending, assigned, in_progress, completed, cancelled]
        description:
          type: string
        location:
          type: string
        createdAt:
          type: string
          format: date-time
    MaintenanceRequestDetail:
      allOf:
        - $ref: '#/components/schemas/MaintenanceRequest'
        - type: object
          properties:
            tenant:
              type: object
              properties:
                fullName:
                  type: string
                email:
                  type: string
                phone:
                  type: string
            property:
              type: object
              properties:
                title:
                  type: string
                address:
                  type: string
            assignedTo:
              type: object
              properties:
                name:
                  type: string
                company:
                  type: string
                phone:
                  type: string
            scheduledDate:
              type: string
              format: date-time
            completedDate:
              type: string
              format: date-time
            resolution:
              type: string
            images:
              type: array
              items:
                type: string
            statusHistory:
              type: array
              items:
                type: object
                properties:
                  status:
                    type: string
                  updatedAt:
                    type: string
                    format: date-time
                  updatedBy:
                    type: string
                  notes:
                    type: string
...
    UpdateMaintenanceRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, assigned, in_progress, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, emergency]
        assignedToId:
          type: string
          format: uuid
        scheduledDate:
          type: string
          format: date-time
        resolution:
          type: string
        notes:
          type: string

    # Lease Template Schemas
...